services/django-api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
mysqlclient>=2.2
redis>=5.0
paho-mqtt>=2.1
gunicorn>=22.0
